refactor(FaceProcessingRouter): Extract all the entry points about face processing to a seperate object
